overhaul of fragment loading.  all fragments now load into the same area.  nested fragments are reduced.  new animation of setting and navigation (things that need larger space hide the passenger controls fragment).

+++ lydia/src/ca/efriesen/lydia/databases/ButtonConfigOpenHelper.java
@@ -11,7 +11,7 @@ import android.util.Log;
 public class ButtonConfigOpenHelper extends SQLiteOpenHelper {
 
 	// If any schema changes are made, the hard coded settings button in the HomeScreenFragment will need updating
-	private static final int DATABASE_VERSION = 8;
+	private static final int DATABASE_VERSION = 11;
 	private static final String DATABASE_NAME = "buttonConfig.db";
 
 	public static final String TABLE_NAME = "button_config";
@@ -19,6 +19,7 @@ public class ButtonConfigOpenHelper extends SQLiteOpenHelper {
 	public static final String COLUMN_ID = "_id";
 	public static final String BUTTONTYPE = "button_type";
 	public static final String DISPLAYAREA = "display_area";
+	public static final String GROUP = "button_group";
 	public static final String POSITION = "position";
 	public static final String TITLE = "title";
 	public static final String ACTION = "action";
@@ -31,13 +32,13 @@ public class ButtonConfigOpenHelper extends SQLiteOpenHelper {
 					" (" + COLUMN_ID + " integer primary key autoincrement, " +
 					BUTTONTYPE + " integer not null, " +
 					DISPLAYAREA + " integer not null, " +
+					GROUP + " integer default 0, " +
 					POSITION + " integer not null, " +
 					TITLE + " text, " +
 					ACTION + " text not null, " +
 					DRAWABLE + " text not null, " +
 					USESDRAWABLE + " integer not null, " +
 					EXTRADATA + " text);";
-
 
 	ButtonConfigOpenHelper(Context context) {
 		super(context, DATABASE_NAME, null, DATABASE_VERSION);
@@ -51,7 +52,12 @@ public class ButtonConfigOpenHelper extends SQLiteOpenHelper {
 	@Override
 	public void onUpgrade(SQLiteDatabase sqLiteDatabase, int oldversion, int newversion) {
 		Log.w(MessageOpenHelper.class.getName(), "Upgrading Database from version " + oldversion + " to " + newversion + ".  This will destroy all data.");
-		sqLiteDatabase.execSQL("DROP TABLE IF EXISTS " + TABLE_NAME);
-		onCreate(sqLiteDatabase);
+		if (oldversion < 8) {
+			sqLiteDatabase.execSQL("DROP TABLE IF EXISTS " + TABLE_NAME);
+			onCreate(sqLiteDatabase);
+		}
+		if (oldversion < 11) {
+			sqLiteDatabase.execSQL("ALTER TABLE " + TABLE_NAME + " ADD COLUMN " + GROUP + " integer DEFAULT 0;");
+		}
 	}
 }
